refactor: Move AvailableEVSELocation to DTO and extract mapping logic
Moved AvailableEVSELocation struct from domain to application layer DTO and extracted data mapping to a new mapper file for better separation of concerns
