+++ internal/application/dto/active_evse_location_dto.go
@@ -1,0 +1,11 @@
+package dto
+
+type AvailableEVSELocation struct {
+	ID        string
+	Name      *string
+	Address   string
+	Latitude  float64
+	Longitude float64
+	UID       string
+	Status    int
+}
